requirements.txt
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
numpy<2.0.0
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.0
transformers==4.33.2
torch==2.0.1
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3
aiohttp==3.8.5
